+++ backend/chainlit/socket.py
@@ -9,6 +9,7 @@ from chainlit.action import Action
 from chainlit.auth import get_current_user, require_login
 from chainlit.config import config
 from chainlit.context import init_ws_context
+from chainlit.chat_context import chat_context
 from chainlit.data import get_data_layer
 from chainlit.element import Element
 from chainlit.logger import logger
@@ -185,6 +186,11 @@ async def connection_successful(sid):
                 {"interaction": "resume", "thread_id": thread.get("id")},
             )
             await config.code.on_chat_resume(thread)
+
+            for step in thread.get("steps", []):
+                if "message" in step["type"]:
+                    chat_context.add(Message.from_dict(step))
+
             await context.emitter.resume_thread(thread)
             return
 
@@ -269,6 +275,36 @@ async def process_message(session: WebsocketSession, payload: MessagePayload):
         ).send()
     finally:
         await context.emitter.task_end()
+
+
+@sio.on("edit_message")
+async def edit_message(sid, payload: MessagePayload):
+    """Handle a message sent by the User."""
+    session = WebsocketSession.require(sid)
+    context = init_ws_context(session)
+
+    messages = chat_context.get()
+
+    orig_message = None
+
+    for message in messages:
+        if orig_message:
+            await message.remove()
+
+        if message.id == payload["message"]["id"]:
+            message.content = payload["message"]["output"]
+            await message.update()
+            orig_message = message
+
+    await context.emitter.task_start()
+
+    if config.code.on_message:
+        try:
+            await config.code.on_message(orig_message)
+        except asyncio.CancelledError:
+            pass
+        finally:
+            await context.emitter.task_end()
 
 
 def is_ui_message_valid(payload: MessagePayload) -> bool: