+++ frontend/src/components/molecules/messages/components/Avatar.tsx
@@ -10,7 +10,7 @@ import {
 } from '@chainlit/react-client';
 
 interface Props {
-  author: string;
+  author?: string;
   hide?: boolean;
 }